requirements.txt
accelerate==0.34.0
av==12.3.0
datasets==2.16.1  
decord==0.6.0  
gradio_client==0.2.9  \\
hf_transfer==0.1.9    \\
hjson==3.1.0  
httpcore==0.17.3  
httpx==0.24.0  
huggingface-hub==0.28.1  
hydra-core==1.3.2  
idna==3.10  
numpy==1.26.1  
omegaconf==2.3.0  
opencv-python==4.11.0.86  
peft==0.4.0  
wandb==0.18.7  \\
transformers==4.45.1
anyio==4.3.0
audioread==3.0.1
av==12.3.0
datasets==2.16.1
einops==0.6.1
einops-exts==0.0.4 \\
ffmpy==0.3.2       \\
fvcore             \\
gradio==4.43.0      \\
gradio_client==1.3.0
ninja==1.11.1.1
notebook==6.4.10      \\
numpy==1.26.1
open_clip_torch==2.26.1   \\
peft==0.4.0
protobuf==4.24.4
pyzmq==25.1.1
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.11.1
sentencepiece==0.1.99
spacy==3.7.1             \\带版本
spacy-legacy==3.0.12
spacy-loggers==1.0.5 
tensorboard==2.9.0
tensorrt==8.6.1           \\
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu121
timm==0.9.11
toml==0.10.2
tqdm==4.66.1
loguru==0.7.3           \\
tenacity                 \\
sqlitedict               \\
evaluate==0.4.3
sacrebleu==2.5.1
pytablewriter==1.2.1
pytest==8.3.5